불필요한 API 호출 방지를 위한 flash / meeting 컴포넌트 분리 (#1044)
