+++ pages/list/index.tsx
@@ -1,15 +1,9 @@
-import { ampli } from '@/ampli';
-import { useQueryMyProfile } from '@api/API_LEGACY/user/hooks';
-import ConfirmModal from '@components/modal/ConfirmModal';
 import CardSkeleton from '@components/page/list/Card/Skeleton';
 import Filter from '@components/page/list/Filter';
 import GridLayout from '@components/page/list/Grid/Layout';
 import { MeetingListOfAll } from '@components/page/list/Grid/List';
 import { SSRSafeSuspense } from '@components/util/SSRSafeSuspense';
-import useModal from '@hooks/useModal';
-import { playgroundLink } from '@sopt-makers/playground-common';
 import type { NextPage } from 'next';
-import { useRouter } from 'next/router';
 import { styled } from 'stitches.config';
 import CrewTab from '@components/CrewTab';
 
@@ -17,19 +11,6 @@ import FloatingButton from '@components/FloatingButton';
 import GuideButton from '@components/GuideButton';
 
 const Home: NextPage = () => {
-  const router = useRouter();
-  const { data: me } = useQueryMyProfile();
-  const { isModalOpened, handleModalOpen, handleModalClose } = useModal();
-
-  const handleMakeMeeting = () => {
-    if (!me?.hasActivities) {
-      handleModalOpen();
-      return;
-    }
-    ampli.clickMakeGroup({ location: router.pathname });
-    router.push('/make');
-  };
-
   return (
     <>
       <div>
@@ -56,16 +37,6 @@ const Home: NextPage = () => {
           <MeetingListOfAll />
         </SSRSafeSuspense>
       </div>
-
-      <ConfirmModal
-        isModalOpened={isModalOpened}
-        message={`모임을 개설하려면\n프로필 작성이 필요해요`}
-        cancelButton="돌아가기"
-        confirmButton="작성하기"
-        handleModalClose={handleModalClose}
-        handleConfirm={() => (window.location.href = `${playgroundLink.memberUpload()}`)}
-      />
-
       <FloatingButton />
     </>
   );
