+++ src/components/page/list/Card/index.tsx
@@ -6,6 +6,7 @@ import { styled } from 'stitches.config';
 import { PART_OPTIONS, PART_VALUES, RECRUITMENT_STATUS } from '@constants/option';
 import { ampli } from '@/ampli';
 import { MeetingListOfFilterResponse } from '@api/API_LEGACY/meeting';
+import DesktopSizeFlashCard from '@components/page/list/Card/DesktopSizeCard/DesktopSizeFlashCard';
 
 interface CardProps {
   bottom?: ReactNode;
@@ -15,6 +16,7 @@ interface CardProps {
 
 function Card({ bottom, meetingData, mobileType }: CardProps) {
   const isAllParts = meetingData.joinableParts?.length === 6 || meetingData.joinableParts === null;
+  const isFlash = meetingData.category === '번쩍';
 
   return (
     <CardWrapper
@@ -31,9 +33,9 @@ function Card({ bottom, meetingData, mobileType }: CardProps) {
         });
       }}
     >
-      <Link href={`/detail?id=${meetingData.id}`}>
+      <Link href={isFlash ? `/detail/flash?id=${meetingData.id}` : `/detail?id=${meetingData.id}`}>
         <DesktopOnly>
-          <DesktopSizeCard meetingData={meetingData} />
+          {isFlash ? <DesktopSizeFlashCard meetingData={meetingData} /> : <DesktopSizeCard meetingData={meetingData} />}
         </DesktopOnly>
         <MobileOnly>
           <MobileSizeCard meetingData={meetingData} isAllParts={isAllParts} mobileType={mobileType} />
